Move payload integration tests from mod.rs to content.rs
